finished module reorganization
